+++ generators/php/src/IR/ConstantDecl.php
@@ -22,6 +22,7 @@ final class ConstantDecl
 	public function __construct(
 		public readonly string $name,
 		public readonly mixed $value,
+		public readonly bool $isLibExport = false,
 	) {
 	}
 }